+++ clojure-ts-mode.el
@@ -1442,17 +1442,38 @@ If NS is defined, then the fully qualified symbol is passed to
                        (seq-sort (lambda (spec1 _spec2)
                                    (equal (car spec1) :block)))))))))
 
-(defun clojure-ts--find-semantic-rules-for-node (node)
-  "Return a list of semantic rules for NODE."
-  (let* ((first-child (clojure-ts--node-child-skip-metadata node 0))
-         (symbol-name (clojure-ts--named-node-text first-child))
-         (symbol-namespace (clojure-ts--node-namespace-text first-child)))
+(defvar-local clojure-ts--dynamic-indent-for-symbol-cache
+  (make-hash-table :test 'equal))
+
+(defvar-local clojure-ts--use-dynamic-indent-cache nil
+  "If set to nil, do not use cache for dynamic indentation rules.")
+
+(defun clojure-ts--find-semantic-rules-for-symbol (node)
+  "Return a list of semantic rules for symbol NODE.
+
+If rules are not found return :not-found symbol."
+  (let ((symbol-name (clojure-ts--named-node-text node))
+        (symbol-namespace (clojure-ts--node-namespace-text node)))
     (or (clojure-ts--dynamic-indent-for-symbol symbol-name symbol-namespace)
         (alist-get symbol-name
                    clojure-ts--semantic-indent-rules-cache
                    nil
                    nil
                    #'equal))))
+
+(defun clojure-ts--find-semantic-rules-for-node (node)
+  "Return a list of semantic rules for NODE."
+  (let ((first-child (clojure-ts--first-value-child node)))
+    (if clojure-ts--use-dynamic-indent-cache
+        ;; `with-memoization' can't cache nil, so a sentinel stands in for "no
+        ;; rules" and is translated back before returning.
+        (let ((rules (with-memoization
+                         (gethash (treesit-node-text first-child)
+                                  clojure-ts--dynamic-indent-for-symbol-cache)
+                       (or (clojure-ts--find-semantic-rules-for-symbol first-child)
+                           :not-found))))
+          (unless (eq rules :not-found) rules))
+      (clojure-ts--find-semantic-rules-for-symbol first-child))))
 
 (defun clojure-ts--find-semantic-rule (node parent current-depth)
   "Return a suitable indentation rule for NODE, considering the CURRENT-DEPTH.
@@ -1464,7 +1485,8 @@ increasing the CURRENT-DEPTH.  If a rule is not found upon reaching the
 root of the syntax tree, it returns nil.  A rule is considered a match
 only if the CURRENT-DEPTH matches the rule's required depth."
   (let* ((idx (- (treesit-node-index node) 2)))
-    (if-let* ((rule-set (clojure-ts--find-semantic-rules-for-node parent)))
+    (if-let* ((rule-set (clojure-ts--find-semantic-rules-for-node parent))
+              ((not (equal rule-set :not-found))))
         (if (zerop current-depth)
             (let ((rule (car rule-set)))
               (if (equal (car rule) :block)
@@ -1899,6 +1921,7 @@ subsequent special arguments based on block indentation rules."
     ;; indentation rules.  First node to skip is the symbol itself.
     (when (equal sexp-type 'cond)
       (if-let* ((rule-set (clojure-ts--find-semantic-rules-for-node node))
+                ((not (equal rule-set :not-found)))
                 (rule (car rule-set))
                 ((equal (car rule) :block)))
           (treesit-beginning-of-thing 'sexp (1- (- (cadr rule))) 'restrict)
@@ -1919,9 +1942,11 @@ between BEG and END."
                            (end (clojure-ts--end-of-defun-pos)))
                        (list start end))))))
   (setq end (copy-marker end))
+  (clrhash clojure-ts--dynamic-indent-for-symbol-cache)
   (let* ((sexps-to-align (clojure-ts--get-nodes-to-align beg (marker-position end)))
          ;; We have to disable it here to avoid endless recursion.
-         (clojure-ts-align-forms-automatically nil))
+         (clojure-ts-align-forms-automatically nil)
+         (clojure-ts--use-dynamic-indent-cache t))
     (save-excursion
       (indent-region beg (marker-position end))
       (dolist (sexp sexps-to-align)